drop dependencies on RescuePrimeDepracated
